+++ dev/com.ibm.ws.concurrent.mp_fat_jakarta/.classpath
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry kind="src" path="fat/src"/>
-	<classpathentry kind="src" path="test-applications/MPContextProp1_3_App/src"/>
-	<classpathentry kind="src" path="test-applications/MPContextProp2_0_App/src"/>
+	<classpathentry kind="src" path="test-applications/MPContextProp1_3_EE9_App/src"/>
+	<classpathentry kind="src" path="test-applications/MPContextProp1_3_EE10_App/src"/>
 	<classpathentry kind="src" path="test-contextproviders/threadprioritycontext/src"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
