+++ dev/com.ibm.ws.concurrent.mp_fat_jakarta/build.gradle
@@ -16,4 +16,3 @@ dependencies {
   requiredLibs project(':com.ibm.tx.jta')
   requiredLibs project(':io.openliberty.org.eclipse.microprofile')
 }
-
