+++ dev/com.ibm.ws.concurrent.mp_fat_jakarta/publish/servers/com.ibm.ws.concurrent.mp.fat.1.3.ee10/server.xml
@@ -1,0 +1,41 @@
+<!--
+    Copyright (c) 2022 IBM Corporation and others.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+   
+    Contributors:
+        IBM Corporation - initial API and implementation
+ -->
+<server>
+
+  <featureManager>
+    <feature>bells-1.0</feature>
+    <feature>cdi-4.0</feature>
+    <feature>componenttest-2.0</feature>
+    <feature>concurrent-3.0</feature>
+    <feature>jndi-1.0</feature>
+    <feature>mpContextPropagation-1.3</feature> 
+    <feature>mpFaultTolerance-4.0</feature>
+    <feature>servlet-6.0</feature>
+  </featureManager>
+
+  <include location="../fatTestPorts.xml"/>
+    
+  <application location="MPContextProp1_3_EE10_App.war"/>
+
+  <!-- For processing of custom thread context providers that are defined in META-INF/services -->
+  <bell>
+    <library id="CustomThreadContextProvidersLib">
+      <file name="${server.config.dir}/lib/customContextProviders2.jar"/>
+    </library>
+  </bell>
+
+  <!-- Needed for application to use a ForkJoinPool, access the thread context class loader, and shut down an unmanaged ExecutorService that the test application creates -->
+  <javaPermission codebase="${server.config.dir}/apps/MPContextProp1_3_EE10_App.war" className="java.lang.RuntimePermission" name="getClassLoader"/>
+  <javaPermission codebase="${server.config.dir}/apps/MPContextProp1_3_EE10_App.war" className="java.lang.RuntimePermission" name="modifyThread"/>
+  <javaPermission codebase="${server.config.dir}/apps/MPContextProp1_3_EE10_App.war" className="java.lang.RuntimePermission" name="setContextClassLoader"/>
+  <javaPermission codebase="${server.config.dir}/apps/MPContextProp1_3_EE10_App.war" className="java.util.PropertyPermission" name="java.util.concurrent.ForkJoinPool.*" actions="read"/>
+
+</server>
