--- dev/com.ibm.ws.concurrent.mp_fat_jakarta/publish/servers/com.ibm.ws.concurrent.mp.fat.2.0/server.xml
@@ -1,48 +1,0 @@
-<!--
-    Copyright (c) 2022 IBM Corporation and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-   
-    Contributors:
-        IBM Corporation - initial API and implementation
- -->
-<server>
-
-  <!-- TODO eventually, this is intended to be an Jakarta EE 10 server 
-       with MicroProfile Context Propagation 2.0.
-       In the mean time, we are using it to test the combination of
-       Concurrency 3.0 (an EE 10 feature) with MP Context Propagation enabled.
-       At some point, Concurrency 3.0 will be prevented from running with EE 9
-       features, and this test will need to be disabled for a time, until
-       the EE 10 compatible version of MP Context Propagation becomes available. -->
-  <featureManager>
-    <feature>bells-1.0</feature>
-    <feature>cdi-3.0</feature> <!-- TODO 4.0 -->
-    <feature>componenttest-2.0</feature>
-    <feature>concurrent-3.0</feature>
-    <feature>jndi-1.0</feature>
-    <feature>mpContextPropagation-1.3</feature> <!-- TODO 2.0 -->
-    <feature>mpFaultTolerance-4.0</feature> <!-- TODO EE 10 compatible version -->
-    <feature>servlet-5.0</feature> <!-- TODO 6.0 -->
-  </featureManager>
-
-  <include location="../fatTestPorts.xml"/>
-    
-  <application location="MPContextProp2_0_App.war"/>
-
-  <!-- For processing of custom thread context providers that are defined in META-INF/services -->
-  <bell>
-    <library id="CustomThreadContextProvidersLib">
-      <file name="${server.config.dir}/lib/customContextProviders2.jar"/>
-    </library>
-  </bell>
-
-  <!-- Needed for application to use a ForkJoinPool, access the thread context class loader, and shut down an unmanaged ExecutorService that the test application creates -->
-  <javaPermission codebase="${server.config.dir}/apps/MPContextProp2_0_App.war" className="java.lang.RuntimePermission" name="getClassLoader"/>
-  <javaPermission codebase="${server.config.dir}/apps/MPContextProp2_0_App.war" className="java.lang.RuntimePermission" name="modifyThread"/>
-  <javaPermission codebase="${server.config.dir}/apps/MPContextProp2_0_App.war" className="java.lang.RuntimePermission" name="setContextClassLoader"/>
-  <javaPermission codebase="${server.config.dir}/apps/MPContextProp2_0_App.war" className="java.util.PropertyPermission" name="java.util.concurrent.ForkJoinPool.*" actions="read"/>
-
-</server>
