+++ dev/com.ibm.ws.concurrent.mp_fat_jakarta/test-applications/MPContextProp1_3_EE10_App/src/concurrent/mp/fat/v13/ee10/web/AsyncClassBean.java
@@ -8,34 +8,26 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package concurrent.mp.fat.v20.web;
+package concurrent.mp.fat.v13.ee10.web;
 
 import java.io.Serializable;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
-import java.util.concurrent.CompletionStage;
 
 import jakarta.enterprise.concurrent.Asynchronous;
-import jakarta.enterprise.concurrent.ManagedExecutorService;
 import jakarta.enterprise.context.ApplicationScoped;
 
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
 
 @ApplicationScoped
-@org.eclipse.microprofile.faulttolerance.Asynchronous
-public class MPFTBean implements Serializable {
+@Asynchronous(executor = "java:comp/eeExecutor") // Should not be allowed at this location
+public class AsyncClassBean implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    @Asynchronous
-    CompletionStage<String> doublyAsync() {
-        return Asynchronous.Result.complete("Should not be able to combine Jakarta Concurrency @Asynchronous on a method " +
-                                            "with MicroProfile Fault Tolerance @Asynchronous on the class");
-    }
-
-    CompletionStage<Object> ftAsyncLookup(String name) {
+    CompletableFuture<Object> asyncLookup(String name) {
         try {
-            ManagedExecutorService executor = (ManagedExecutorService) InitialContext.doLookup(name);
-            return executor.completedFuture(executor);
+            return Asynchronous.Result.complete(InitialContext.doLookup(name));
         } catch (NamingException x) {
             throw new CompletionException(x);
         }
