+++ dev/com.ibm.ws.concurrent.mp_fat_jakarta/test-applications/MPContextProp1_3_EE10_App/src/concurrent/mp/fat/v13/ee10/web/MPAppBean.java
@@ -8,11 +8,12 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package concurrent.mp.fat.v20.web;
+package concurrent.mp.fat.v13.ee10.web;
 
 import java.io.Serializable;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
+import java.util.concurrent.CompletionStage;
 
 import jakarta.enterprise.concurrent.Asynchronous;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -21,11 +22,26 @@ import javax.naming.InitialContext;
 import javax.naming.NamingException;
 
 @ApplicationScoped
-@Asynchronous(executor = "java:comp/eeExecutor") // Should not be allowed at this location
-public class AsyncClassBean implements Serializable {
+public class MPAppBean implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    CompletableFuture<Object> asyncLookup(String name) {
+    @Asynchronous(executor = "java:comp/eeExecutor")
+    CompletableFuture<Object> eeAsyncLookup(String name) {
+        try {
+            return Asynchronous.Result.complete(InitialContext.doLookup(name));
+        } catch (NamingException x) {
+            throw new CompletionException(x);
+        }
+    }
+
+    @Asynchronous
+    @org.eclipse.microprofile.faulttolerance.Asynchronous
+    CompletionStage<String> doublyAsync() {
+        return Asynchronous.Result.complete("Should not be able to combine different @Asynchronous annotations on a method");
+    }
+
+    @Asynchronous(executor = "java:module/env/defaultExecutorRef")
+    CompletableFuture<Object> mpAsyncLookup(String name) {
         try {
             return Asynchronous.Result.complete(InitialContext.doLookup(name));
         } catch (NamingException x) {
